Sketch 2: more classy... idk if this fits the codebase more
